+++ opensyde_tool/src/system_definition/node_edit/C_SdNdeIpAddressConfigurationWidget.cpp
@@ -165,8 +165,7 @@ void C_SdNdeIpAddressConfigurationWidget::keyPressEvent(QKeyEvent * const opc_Ke
    bool q_CallOrg = true;
 
    //Handle all enter key cases manually
-   if ((opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Enter)) ||
-       (opc_KeyEvent->key() == static_cast<int32_t>(Qt::Key_Return)))
+   if ((opc_KeyEvent->key() == Qt::Key_Enter) || (opc_KeyEvent->key() == Qt::Key_Return))
    {
       if (((opc_KeyEvent->modifiers().testFlag(Qt::ControlModifier) == true) &&
            (opc_KeyEvent->modifiers().testFlag(Qt::AltModifier) == false)) &&
